# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: eclEZapp API
  version: 0.1.0
  description: Contrato inicial da API compartilhada pelas aplicações web e mobile.
servers:
  - url: http://localhost:8080
    description: Ambiente local
paths:
...
  /api/v1/parishes/{parishId}/servants:
    parameters:
      - $ref: '#/components/parameters/ParishId'
    get:
      operationId: listServants
      summary: Lista servos da paróquia ativa para padre ou administrador
      security:
        - sessionCookie: []
      parameters:
        - in: query
          name: search
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Lista paginada de servos
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data:
                    type: array
                    items: { $ref: '#/components/schemas/Servant' }
                  meta: { $ref: '#/components/schemas/ResponseMeta' }
        '401': { description: Não autenticado }
        '403': { description: Sem contexto ou papel autorizado }
    post:
      operationId: createServant
      summary: Cria pessoa e vínculo de servo sem gerar usuário
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [full_name]
              properties:
                full_name: { type: string, minLength: 2, maxLength: 180 }
                preferred_name: { type: [string, 'null'], maxLength: 100 }
                phone: { type: [string, 'null'], maxLength: 32 }
                email: { type: [string, 'null'], format: email }
                joined_on: { type: [string, 'null'], format: date }
      responses:
        '201':
          description: Servo criado sem credenciais
          content:
            application/json:
              schema:
                type: object
                required: [data, meta]
                properties:
                  data: { $ref: '#/components/schemas/Servant' }
                  meta: { $ref: '#/components/schemas/ResponseMeta' }
        '401': { description: Não autenticado }
        '403': { description: Sem contexto ou papel autorizado }
        '422': { description: Dados inválidos }
...
  /api/v1/parishes/{parishId}/servants/{servantId}/functions:
    parameters:
      - $ref: '#/components/parameters/ParishId'
      - in: path
        name: servantId
        required: true
        schema: { type: string, format: uuid }
    post:
      operationId: qualifyServant
      summary: Cria ou atualiza a habilitacao pastoral de um servo
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [pastoral_function_id]
              properties:
                pastoral_function_id: { type: string, format: uuid }
                status: { type: string, enum: [PENDING, QUALIFIED, SUSPENDED, EXPIRED] }
                qualified_on: { type: [string, 'null'], format: date }
                expires_on: { type: [string, 'null'], format: date }
                notes: { type: [string, 'null'], maxLength: 2000 }
      responses:
        '200': { description: Habilitacao existente atualizada }
        '201': { description: Habilitacao criada }
        '401': { description: Nao autenticado }
        '403': { description: Sem contexto ou papel autorizado }
        '404': { description: Servo nao encontrado na paroquia ativa }
        '422': { description: 'Funcao externa, inativa ou dados invalidos' }
components:
  parameters:
    ParishId:
      in: path
      name: parishId
      required: true
      schema: { type: string, format: uuid }
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: laravel_session
  schemas:
    ResponseMeta:
      type: object
      required: [request_id]
      properties:
        request_id:
          type: string
...
    Servant:
      type: object
      required: [id, parish_id, person, status, joined_on, left_on, has_user, functions]
      properties:
        id: { type: string, format: uuid }
        parish_id: { type: string, format: uuid }
        person:
          type: object
          required: [id, full_name, preferred_name, phone, email]
          properties:
            id: { type: string, format: uuid }
            full_name: { type: string }
            preferred_name: { type: [string, 'null'] }
            phone: { type: [string, 'null'] }
            email: { type: [string, 'null'], format: email }
        status: { type: string, enum: [ACTIVE, INACTIVE, SUSPENDED] }
        joined_on: { type: [string, 'null'], format: date }
        left_on: { type: [string, 'null'], format: date }
        has_user: { type: boolean }
        functions:
          type: array
          items:
            type: object
            required: [id, status, function_id, function_name, area_name]
            properties:
              id: { type: string, format: uuid }
              status: { type: string, enum: [PENDING, QUALIFIED, SUSPENDED, EXPIRED] }
              function_id: { type: string, format: uuid }
              function_name: { type: string }
              area_name: { type: [string, 'null'] }
...